+++ src/helper/type-manager.ts
@@ -38,7 +38,8 @@ import {
   removeOuterContextPrefixInNamespace
 } from './utils';
 
-const DEFAULT_CUSTOM_FUNCTION_DESCRIPTION = `This is a custom method. You can add a description for this method by adding a comment above or after the function.
+const DEFAULT_CUSTOM_FUNCTION_DESCRIPTION =
+  `This is a custom method. You can add a description for this method by adding a comment above or after the function.
 \`\`\`
 myFunction = function(a, b, c) // This function does xyz
 \`\`\`
@@ -346,7 +347,10 @@ export class TypeMap {
     item: ASTFunctionStatement
   ): TypeInfoWithDefinition | null {
     const me = this;
-    const description = me.getItemDescription(item, DEFAULT_CUSTOM_FUNCTION_DESCRIPTION);
+    const description = me.getItemDescription(
+      item,
+      DEFAULT_CUSTOM_FUNCTION_DESCRIPTION
+    );
 
     return new TypeInfoWithDefinition('anonymous', ['function'], {
       arguments: item.parameters.map((arg: ASTBase) => {